requirements.txt
-e .
colour-runner==0.0.5
coverage==4.3.4
dj-database-url==0.4.2
dj-inmemorystorage==1.4.0
django-crispy-forms==1.6.1
factory-boy==2.8.1
flake8-import-order==0.11
flake8==3.2.1
incuna-test-utils==6.6.0
psycopg2==2.6.2
